checkin of current jenkins jobs, now beautified via libXML
The job configurations are now beautified by default in order to
have meaningful diffs and to prevent useless diffs, as the original
jenkins xml looks slightly different than the libXML output.

We now use the libXML output to be able to edit the job configs
with libXML via scripts (trackupstream is the first example) and
still have reasonable diffs.

See commit message of commit 5088a4 for details about the new workflow.