crowbar-prep.sh: allow overriding of defaults
This allows people to put stuff like

  export HOST_MEDIA_MIRROR_DEFAULT=/mnt
  export HOST_MIRROR_DEFAULT=/srv/mirrors

in their shell profile, and then still be able to override via the
command-line.  Note that due to the cat crowbar-prep.sh | ssh ...
style of invocation, the export would have to go in the shell profile
on the admin node, not the node where crowbar-prep.sh is available.
So it may still be easier to use the -m and -n options.