mkcloud: Add 9p fs for local repos to admin node
Add $localreposdir_src var and if available add 9p filesystem to admin node
so in the future local repositories can be used for the admin node.
This is still WIP but with this commit at least the 9p fs is available
in the VM and the var $localreposdir_target is exported for
the sshrun of qa_crowbarsetup.sh.