qa_crowbarsetup: check if testvm ssh is up actively
This patch changes the way qa_crowbar waits until `testvm` SSH keys are
ready to log in. Originally the script was sleeping for 200 secs and then
just connecting, now the script checks 40 times every 5 seconds to know if
the keys are ready to be used.

`ssh-keygen -R $vmip` is being used, because the server's fingerprint changes
a few seconds after `sshd` started listening connections for the first time.