Remove adding of local repos to the admin node in TESTHEAD setups
Clarify documentation that TESTHEAD picks up the media from
Staging and tests it against the update-testing repositories.

The extra repos like Devel_Cloud_Shared and the like were only
breaking the run most of the time, and we should really test the
media only - otherwise packages might be available that have
been forgotten from the media.