create-vm: Add extra options for nic and disk
Add parameters to define extra options for network interface and disk.
This is needed to setup a VM which is later used to boot a Xen kernel.
Xen does not understand virtio so to use the VM to boot a Xen kernel,
add e.g. "--disk-opts bus=sata --nic-opts model=e1000".