mkcloud: don't run /etc/init.d/boot.local
Instead of appending a large chunk of code to boot.local and then
executing the whole file, write it to a separate file which is then
executed and also invoked from boot.local on boot.  This avoids any
unpredictable side-effects of running boot.local after the system is
fully booted up, and also allows us to cleanly migrate to new behaviour
in future on systems which have already had mkcloud run on them, simply
by replacing boot.mkcloud with a new version.