qa_crowbarsetup.sh: break some overly long lines
Conventional wisdom across the industry is that it is a bad idea to
exceed 80 column lines. This is visible in countless areas, from the
Linux kernel coding style through to PEP8 and beyond.  Long lines
definitely make code harder to read for some people, so even if you
don't *personally* feel it increases readability for you, please still
be considerate of other people.

It's not just a matter of taste either.  Some user interfaces do not
automatically wrap lines, so in those cases long lines can cause
significant usability issues.  For example, when viewing diffs on parts
of github, the lack of line wrapping means that you have to use a
horizontal scroll bar to see the whole line.  This is a major PITA.

In the case where there is an `if` statement with a long conditional, it
means that the conditional gets broken across multiple lines.  In this
case it's more readable if the `then` is on its own line, so you can
clearly see where the conditional ends and the body begins.  This type
of approach is advocated in

  http://wiki.bash-hackers.org/scripting/style

This requires ignoring a ridiculous bashate rule (E011) which complains
if 'then' is not on the same line as the 'if' beginning a conditional.