mkcloud: allow independent creation/use of DRBD disks
Currently the hacloud option is very opinionated and requires at
least 11 VMs to work.  On a machine with only 16GB of RAM this
is totally useless.  There are plans to fix this:

  https://trello.com/c/9viz41KH

However as a short-term workaround it is good enough to simply
allow DRBD disks to be set up even if the hacloud parameter is
not set to 1; then we can do:

  mkcloud instonly

and then feed a pre-prepared YAML to crowbar batch to set up
the cloud.