mkcloud: watch out for incomplete LVM device filter
mkcloud creates LVs for use as DRBD data disks, which are then used by
the controller VMs as PVs to build drbd VGs on top of.  However we need
to prevent the host from detecting the LVs inside the resulting drbd
VGs, because they only concern the controller VM guests.  This requires
tweaking lvm.conf which is not easy to automate, but we can at least
check whether it was done right, and bail if not.