mkcloud: allow cloud instances to get responses from dnsmasq
The problem is that we generate DNS queries from public IPs
routed over the public VLAN and dnsmasq did only listen
on the admin interface, so could not respond.

libvirt autodetects and uses the bind-dynamic dnsmasq feature
in src/network/bridge_driver.c and src/util/virdnsmasq.c
by looking for the bind-dynamic string in the dnsmasq help
and we can prevent it by replacing it there.