mkcloud: allow cloud instances to get responses from dnsmasq
The problem is that we generate DNS queries from public IPs
routed over the public VLAN and dnsmasq did only listen
on the admin interface, so could not respond.

libvirt autodetects and uses the bind-dynamic dnsmasq feature
in src/network/bridge_driver.c and src/util/virdnsmasq.c
by looking for the bind-dynamic string in the dnsmasq help
and we can prevent it by replacing it there.

Revert "mkcloud: allow cloud instances to get responses from dnsmasq"

This reverts commit 30ab35315ff01b7def2b578b53376c55f5ae74c1.

The problem with emitting an error was that this needed manual intervention
on every new mkcloud host and after every libvirt upgrade
while the mkcloud user expects this to just work.