put top-level code in main() function
This draws a clear line between UI-related code and non-UI code, which
will also make it easier in the future if we ever want to use this as a
module as well as a script.

This is idiomatic Python, e.g.

- http://stackoverflow.com/a/20158605/179332
- https://google-styleguide.googlecode.com/svn/trunk/pyguide.html#Main
- https://www.chromium.org/chromium-os/python-style-guidelines#TOC-Main-Modules

and the vast majority of OpenStack executables adopt the same
convention, thanks to pbr.