clarify use of non-clustered nodes
Following on from f9938fbf, the 'nodescloud' variable actually
referred only to nodes not in a cluster, so rename this to the
more explicit 'unclustered_nodes'.  Also make it clear in the
node allocation summary that these unclustered nodes are available
for use not just as compute (i.e. hypervisor) nodes, but also
potentially for (Ceph) storage.