github-status: allow to filter pr lists by branch name
With the new parameter the PRs can be filtered by branch name.
This is needed for gate testing the crowbar PRs.

After the crowbar (repository) merge the branch name to be monitored
for PRs differs between the old and new repos. This new filter will
be used in jenkins to get a list of only relevant PRs per repo.