qa_crowbarsetup: Use NFS for /var/lib/glance/images when doing HA
The most common error when using HA is to use the file backend in glance
without sharing /var/lib/glance/images. This was also done in mkcloud.

To fix this, we create a NFS share on the admin server dedicated to
this, and we have a nfs_client proposal for the nodes that are on the
services cluster.