qa_openstack: Skip Manila tempest tests
In Liberty, Manila uses the tempest plugin interface to add tests to tempest.
So tempest now runs the Manila tests. The problem is that we need a image
(with nfs and samba installed and some other modifications) to be able
to run the tests.
Skip the tests for now until we have a image available.