qa_crowbarsetup: Prepare Manila generic driver for testing
Manila is integrated into Cloud 6 and we want to test it. Upstream
uses the so called "generic driver" as backend and runs tempest tests.
To be able to run the tempest tests we need to prepare the setup.
The generic driver uses a cloud instance as a share server. The instance
mounts a cinder block device, adds a FS on top and uses the common
tools (nfs-server, samba) to export shares.
To manage the share-server, the manila-share service (which may run on
multiple compute nodes) need to acccess the share-server via ssh.
So we basically need:
- a new network which is available on the crowbar nodes
- a neutron provider network to access the nodes in the crowbar network
- a nova instance which has access to the new neutron network

In addition to that, the image used for the nova instance must have some
software preinstalled (i.e. nfs-server, samba). Also some security-group
rules are needed to allow NFS/CIFS traffic between the share server
instance and the consumers (other instances in the tenant).