Support for more than one controller VM creation
In HA deployments, with more than one controller node, only the first
controller controller VM was assigned an increased memory footprint. The
other controller VM's got the normal amount of memory leading to some
out-of-memory errors. This implementation assigns the same amount of
(increased) memory to all controller VMs.

The former implementation already took care of several controller nodes
in HA deployments. The number of controller nodes could be specified
with the clusterconfig:services configuration variable. This
implementation doesn't change the behavior of that variable.