Avoid node update races on allocation
Horrible workaround for crowbar bug. Crowbar first sets
the node state to "discovered", and then goes through and
calls the transition hooks over all barclamps, which can take
quite a while. However, if we move forward already we're causing
tons of stress on crowbar with the node rename, and then it will
just run into races. Actual fix would be to only change the
state once the state is actually changed. (d'oh!)