xen: Add more memory to Xen nodes
When using Xen, dom0 uses half of the memory and with our current xen job,
nova-compute can not be started due to missing memory.
This leads to a ComputerFilter error in the nova scheduler and no
manila-service VM or other VMs during testing can be started.
So add more memory to the nodes.