Fix handling of manila service instance during upgrade
During the upgrade the manila service vm is shutdown. So when before
re-applying the manila barclamp check for the state of the manial-service
instance. If it exists and is in state "SHUTOFF" restart it. If it does not
exist boot a new instance. Otherwise assume it's running.

Additional fix: Set OS_TENANT_NAME to "manila-service" also if the Project
already exists.