Fix for rebootcloud
When the cloud is rebooted, the nodes run crowbar_join, which runs ntpdate to
update the nodes' date. Unfortunately the ntpd instance on the crowbar node
takes a while before it considers itself trustworthy enough to give out
temporal advice. This means that the nodes will hang on cowbar_join until
crowbar's ntpd finally pulls itself together. Sometimes ntpd takes longer than
the nodes are willing to wait, and the impatient nodes will not be considered
`ready` by crowbar, i.e. the cloud is not fully up. This patch adds an
additional wait after rebooting the crowbar node.