scripts: mkcloud: Set correct MSS for outgoing traffic
The default 1500 MSS for vNICs may not be the best choice if the host
is using a lower value for its external interface. Moreover it's
probably wrong since 1500 is usually the NIC's MTU so the MSS needs to
be slighly lower (usually 40 bytes to account for the IP and TCP
headers).
This could lead to networking issues when the VMs try to access the
external network. It's best to set the MSS value to match the lowest
host's MTU less 40 bytes for the TCP and IP headers in order to make
sure the nodes can reach the network especially during the mkcloud
phases where fetching and installing from repositories is fairly common.
Finally, this patch takes care of removing stale iptables rules due to
previous deployments that did not clean up properly.