fix scenariofile logic and help text
cee28264 from #1103 introduced faulty logic and misleading help text.

If $scenariofile is set by the user, then $scenario is totally
ignored. This is because

    : ${foo:=bar}

will not set $foo if it was already set.

So if $scenariofile is set, it makes more sense to complain (or at least
warn) if $scenario is set (since it would be ignored) rather than if
it's unset.

If that was confusing, maybe this will help:

| scenariofile | scenario | old behaviour          | new behaviour
| ------------ | -------- | ---------------------- | -----------------
|    not set   |  not set | correct behaviour      | no change
|    not set   |    set   | correct behaviour      | no change
|      set     |  not set | incorrectly prohibited | allowed
|      set     |    set   | ignores $scenario      | warn