fix scenariofile logic and help text
cee28264 from #1103 introduced faulty logic and misleading help text.

If $scenariofile is set by the user, then $scenario is totally
ignored if set. This is because

    : ${foo:=bar}

will not set $foo if it was already set.  And if $scenario isn't set
then the code bails for no good reason.

If that was confusing, maybe this will help:

| scenariofile | scenario | behaviour
| ------------ | -------- | ----------------------
|    not set   |  not set | correct behaviour
|    not set   |    set   | correct behaviour
|      set     |  not set | incorrectly prohibited
|      set     |    set   | ignores $scenario

Additionally, the names "scenario" and "scenariofile" are vague and it's
not intuitive what is the difference between them.

So instead we replace the confusing $scenariofile parameter with
$scenario_dir, and that gets joined with $scenario to obtain
$scenario_path which points to the file to scp to the admin node.