mkcloud: Add ostestroptions option
Add an extra option called "ostestroptions". When this option is set,
in the testsetup step ostestr is called (after the usual tempest run)
with the given parameters.
This is useful when the tempest run only executes the smoke tests but
additional tests need to be executed.
For example, Magnum has no smoke tests but for a Magnum CI job we want
to execute the usual tempest smoke test run plus some (or all) tests
for Magnum. This is now possible with:

export ostestroptions=" --regex '^magnum.tests.functional.api'"