Rework the automatic selection of storage nodes
Previously we always picked the 2nd and 3rd node to deploy ceph on (in cloud 7).
With the switch of Cloud to SP2 this doesn't work anymore in an HA case.  The
2nd node will also get a pacemaker role assinged and we get conflicting OS
requirements for that node.

This change now takes the cluster assignment into account and only assigns the
storage roles to nodes that are not part of any cluster (which is a bad choice
regardless of the OS conflict anyway).