Avoid using "knife node from file ..."
Update the node objects in place now using "knife node exec".

"knife node from file" causes the automatic attributes of the node to be
dropped from the database (until the next chef-client run on that node).
This causes problems when executed on unallocated nodes, as done in
"cluster_nodes_assignment" now.