Fix storage method for magnum job
want_ceph shouldn't be used anymore in Jenkins (you can still use it when you
run mkcloud locally). But Jenkins has a variable called "storage_method" which
is implemented in the openstack-mkcloud trigger job. And depending on the value
this variable overrides want_ceph and want_swift when the openstack-mkcloud
trigger is used.