support multiple non-routable clouds per host
Allow DNAT-based port-forwarding to multiple clouds on the same host.

This is mainly for blacher.arch.suse.de at the moment, since all the
mkch*.cloud.suse.de hosts have routable cloud networks, so DNAT is not
required, and any clouds on personal workstations / laptops can of
course be accessed directly from the machine's console.

The expectation is that hosts like blacher will set cloud_port_offset to
1000 + 100 * $cloud_number, i.e. the first cloud will have the same
offset of 1100 as before, but higher clouds will have offsets 1200,
1300, ... etc.

Note that the offsets could not be multiples of 1000, because this would
cause conflicts of the same DNAT port being used for multiple clouds,
since the Crowbar port 3000 and Chef port 4000 are also 1000 apart.