Fix selection of nodes for cinder-volume, ceilometer-agent and nova-compute
Commit 45a66ee90 broke HA deployments on releases other than Cloud 7
that had more than one possible node for cinder-volume, ceilometer-agent
or nova-compute. This is caused by not breaking $unclustered_nodes into
an array as expected by the rest of the code in custom_configuration().

Additionally, we add some quoting for the -z tests to avoid whitespace
issues that triggered "[: too many arguments" errors in the past.