Breakdown gating jobs into separate entities
Until now, gating jobs are based on three main tests: 2 nodes, 4 nodes
and HA. The HA job runs every night but not the other two. Separating
those tasks into their own jobs will make the run nightly.

The main purpose is that when there's a gating failure you are not
certain if the problem was already there or was triggered in the gate.
Now we will have that information.