squashed commits - only for testrun
mkcloud: libvirt nodesetup restructuring and cleanup

The new function to create libvirt configs is now being
used for regular nodes as well as lonely nodes.
Background is also support DRBD volumes on nodes 1 and 2
even if they are lonely nodes. This is required for an HA
setup consisting of lonely nodes only.