mkcloud-dirmaint: Prepare for deploying muliple enviroments
This removes some hardcoded things from the dirmaint driver. And adds
code to "_cleanup" and "_prepare" to allow for multiple clouds being
deployed from a single host.

It will now setup automatically set a vswitch for the cloud, plug a nic
of the host into it and configure dnsmasq for the respective network.

However this code currently assumes that $cloud on z/VM will always
match "mkcl<single-hex-digit>" as it derives the mac addresses and vdev
names from that <single-hex-digit>.