set storage_method to swift (bsc#1001193)
due to the number of nodes in the ha_compute job ceph gets used
as a storage_method.
this is problematic on cloud7 as it will set the compute nodes to
be sles12-sp1. but sles12-sp2 is needed for HA by Crowbar, as there
are no more HA repos in the cloud-repos.yml file under the sles12-sp1
platform