syncSLErepos: Sync latest SUSE namespace version of SES
While SES is still in development we want to have the latest state
from the SUSE namespace. This should help with our current CI problem
where the package version of ceph is different to the version
"ceph --version" prints out. That problem is fixed on IBS in the
SUSE namespace but not yet in the currently used source for SES sync.