Try installing all available revisions
This might be a workaround - the situation is that on the SDK, we had a
newer version of libopenssl-devel, but there was no openssl included
with the same revision, so installing libopenss-devel failed. To
workaround this issue, we are querying all the available revisions of a
specific package and trying to install them starting from the latest.