jjb: shorter github status context, separate testbuild status
The status message of the github status in our PRs was always
truncated. In order to make more room for the message we shorten
the context string:
  ci.suse.de/openstack-mkcloud/gating => suse/mkcloud

Additionally the messages are made consistent for all similar jobs.

Furthermore the testbuild jobs (preliminary job for a PR gating) now
have their own status to better reflect that two steps are needed and
to show both states and have links to both CI jobs.
A testbuild job now needs these status results:
  suse/mkcloud/testbuild  (building the PR RPM package)
  suse/mkcloud            (the PR gating job)