mkcloud: move qa_crowbarsetup help to separate file
The goal is to untie mkcloud and qa_crowbarsetup.
So we removed the sourcing of qa_crowbarsetup a
few weeks ago (this is a feature).
Breakages should not be fixed by sourcing again.

The idea behind this is, that mkcloud can change more
radical than qa_crowbarsetup. The latter will most likely
remain a collection of shell functions. But mkcloud
will need to change sooner or later into a OOP style
(this will not be bash).