Save cloudsource variable to avoid override
The cloudsource variable is changed during the upgrade process. It is
updated from the upgrade_cloudsource variable. That update is later
overriden during the call to set_proposalvars, where cloudsource is read
from disk (/etc/cloudsource). In those places where we change
cloudsource now we also write it to disk.