qa_openstack: Deploy more services
When running qa_openstack, also deploy Sahara, Designate and
Gnocchi.

Note: All theses services are not yet fully configured and started
in openstack-quickstart. But this commit still helps with manual testing
because the database and endpoints are created and with that it is easier
to continue with enablement in cleanvm.