match localreposdir_src with clouddata
localreposdir_src now should point to a hierarchy that looks like
clouddata. This includes both SLES and Cloud directories. This makes it
possible to simply rsync some folders from clouddata, and point to that
directory with localreposdir_src. Please note that it will change the
assumed layout to where localreposdir_src points to.