Allow vendor change on nodes when upgrading packages from develcloud6
When the origin (Cloud6) and target (Cloud7) sources have different
vendor, it is not possible to non-interactively upgrade set of packages
at compute nodes, which is needed for non-disruptive upgrade.

It works correctly with official repositories, but they are not
up to date during development. So for now, we tell zypper to ignore
vendor change.