jtsync: include new status in notification
It's much more helpful if the new status is immediately visible via a
notification (which can appear in many forms, e.g. via an email from
Trello, or as an notification from the Android app), rather than
having to load the Trello app or website in order to see the new
status.