automation: Use generic uefi firmware
Instead of using a uefi firmware with secure boot
enabled and keys in it, use a generic one
with no keys and secure boot disabled.

Code is still in there in case we start signing the
uefi bootloader in the future and want to use secure
boot to confirm that we are signing the images
correctly.

Also disables 2 tests for the livbirt_setup.py
library that were failing before until we have
a fix for them