keystone: Test LDAP configuration on cloud 7
Domain-specific identity backends can be configured to enable the same
functionality as the hybrid backend, so let's start testing it.

LDAP users are now in their own domain so we have to specify that domain
when looking up users. Moreover, keystone builds its own ID mapping for
LDAP users so their suseid is no longer a visible attribute from
openstack.