jjb: allow to disable projects
Templates set the variable disabled=true to disable
an entire project. The jobs are still there and deployed.

This can help when ressouces are low or the workers of an
additional arch are not available - they can be disabled now
with a small change to the project configuration in jjb.