Use ldappool with keystone LDAP backend
ldappool significantly speeds up LDAP requests[1] so it's a good idea to use
it. We weren't using it before because there was a bug making it incompatible
with the use_ssl parameter[2]. The bug has been fixed but not released, but we
can use LDAPS rather than StartTLS and avoid the problem entirely.

[1] https://www.mattfischer.com/blog/?p=624
[2] https://review.openstack.org/#/c/443264/