add debug_mkcloud job parameter
mkcloud does set -x and configures PS4 for tracing if debug_mkcloud is
set.  However we were already calling mkcloud via bash -x which
enforced set -x regardless of this variable, and there was no way to
set it via openstack-mkcloud.  So fix both of these issues by making
it configurable in the job, and default to 1 to be consistent with the
default for debug_qa_crowbarsetup.