mkcloud: Stop chef-client on nodes after a failure
If the mkcloud environment runs into a failure, stop the periodic chef-client
run on all nodes.
This is useful to not alter the environment where the failure happens. Sometimes
there is a failure and the periodic chef-client run fixes the failure afterwards
which makes it more difficult to debug problems in the failed environment.