prevent irrelevant 404 errors from missing robots.txt
We don't define robots.txt in these cases, let alone rely on it,
and if wget searches for a non-existent one, it outputs:

    Loading robots.txt; please ignore errors.
    --2017-03-26 23:03:11--  http://download.nue.suse.com/robots.txt
    Reusing existing connection to download.nue.suse.com:80.
    HTTP request sent, awaiting response... 404 Not Found
    2017-03-26 23:03:11 ERROR 404: Not Found.

the last line of which generates a false positive in the Jenkins Log
Parser Plugin.  So we simply disable this mode in these particular
cases to eliminate the errors.