Raise the nodenumber so the default swift replica count is correct
It seems that due to a very old version of swift in cloud6
we cannot change the default number of replicas so it will
always default to 3. Using a nodenumber count of 3 makes the
default number of nodes assigned to the swift-storage role to be
2 which doesnt sync with the required number of replicas, causing
the upgrade prechecks to fail.
Setting the nodenumber to 4 makes the default number of swift-storage
nodes to be 3, making the upgrade prechecks pass.