avoids overly alarmist errors when looking up domains
4b9bc232c4adf1913c5bfd8e8951124e970e192b tried to address the problem
of the vm-start Python script causing libvirt spewing out error
messages like

    libvirt: QEMU Driver error : Domain not found: no domain with matching name

when the existence of a domain is checked in order to perform
cleanup prior to starting it.  This was important for avoiding false
positives in the Jenkins Log Parser plugin.

However unfortunately the approach used had a few issues:

- It missed one of the places vm-start is called.

- Creating a pipeline with grep interferes with the exit status
  returned by the vm-start script, and replaces it with the exit
  status from the grep, which we don't care about.

- Working around this exit status issue using $PIPESTATUS is tricky,
  very ugly, and doesn't protect against calls to vm-start being added
  elsewhere without the same protection.

So instead we tackle the problem at source, by fixing the
libvirt_setup.py library to look up domains in a way which doesn't
generate errors.  Annoyingly, the libvirt API doesn't lift a finger to
help accomplish this, so we have to write our own wrapper around
libvirt.listAllDomains() to achieve it.

At the same time, we refactor all calls to vm-start to go through a
newly extracted libvirt_vm_start() function, to avoid duplicating
logic like the use of sudo and the script's location.  This will also
help in the future if we need to change the behaviour of all
invocations of vm-start.