Set timezone for crowbar admin node
The jenkins build hosts use UTC, and the nodes that crowbar deploys are
configured to use UTC with autoyast, but the crowbar admin node itself
is set to the local timezone, CEST, because there is nothing in crowbar
that sets it otherwise. This makes sense because in our documentation
we start the process off with installing the base operating system
manually, which is where the operator would normally configure the
timezone. Having inconsistent timezones between the nodes causes a lot
of pain when debugging failed jobs because the log files and the output
of 'date' do not sync up. This patch explicitly sets the timezone on
the admin node during the prepareinstcrowbar phase.