crowbar-testbuild: rewrite in ruby, fully configurable
The redeployment of our CI on the SAP side requires our tools
to be more flexible and configurable to allow them to be run in
different environments.

As making crowbar-testbuild configurable needed to touch almost
every line of the code.
Thus it was a good opportunity for a clean rewrite in ruby.