ignore output of "git show" when testing PRs
When openstack-mkcloud is testing PRs, it does a

    git --no-pager show

to show the changes in the PR being tested.  However if the changes,
or even the context around the changes, mentions something like
"error" or "fail" or "fatal" then the Log Parser plugin will
incorrectly flag this as a false positive.  So we indent the "git
show" output by a unique prefix and then configure the plugin to
ignore anything with that prefix.