mkcloud-common: introduce variables for nfs, repo, rsync and smt server
The goal of this is to drop the $clouddata variable.
clouddata was a host at suse which offered many services in parallel.
We now need to support to configure the services individually and
therefore need separate variables.

The idea is:
 $reposerver, $nfsserver, $rsyncserver and $smtserver are only set from outside
They can be set to an ip or a hostname. But these variables must not
be used in the script(s).
The only ones that are to be used are the ones with the suffixes:
'*_ip' or '*_fqdn'

So this means:
The user sets
  export reposerver="somehostname.company.ldt"
mkcloud-common.sh will resolve to suffix variables.
In the code we only use:
  wget $reposerver_fqdn/some/path/package.rpm