ceph: Fix radosgw s3 connection for SSL
When SSL is turned on, the RadosGW server uses port 8081[1][2] and does
not listen on port 8080. Hard-coding 8080 as the connection port was
causing the RadosGW tests to fail when SSL was turned on[3]. This patch
makes the port number and is_secure flag conditional on whether SSL is
being used so that the test bucket can successfully be created even in
an SSL environment.

[1] https://github.com/crowbar/crowbar-ceph/blob/d290aaca5cddd4539e512355b26efe9ddbe99397/chef/cookbooks/ceph/recipes/conf.rb#L48
[2] https://github.com/crowbar/crowbar-ceph/blob/2bce1372a2e6aba8570d0f30f3cab88839d09d5c/chef/cookbooks/ceph/attributes/radosgw.rb#L22
[3] https://ci.suse.de/job/openstack-mkcloud/61042/