qa_crowbarsetup: drop dead code for pacemaker configuration
Addendum for commit message of ea43af4:
When we switched to stonith mode "libvirt" in commit ea43af4,
this code block was disabled by wrapping it in an "if" conditional
which was hardcoded to always return false. As a result we have
dead code here which just reduces readability.

So get rid of the dead code now.