ha: Really abort when crm failures are detected
Currently, when an error occurs during a remote command, the exit code
from the ssh call is captured and returned without affecting the
deployment. We really do want the deployment to stop when errors are
detected, so this patch wraps the crm checks with 'safely' so that
errors cause the run to abort and supportconfigs collected.