ha: Fix check_crm_failcounts function
This function is run with the "safely" wrapper so that if it fails while
running on a remote host, it causes the run to abort. This is a problem
because the function works by checking the output of grep, which when
finding no matches (which is a success in this case), exits 1. Since the
function apparently fails, the "safely" wrapper causes the run to abort
even though everything is fine. This patch ensures the function returns
a success code if it made it past the "complain", which would cause the
function to exit before making it to the end.