Set mtu of public/nova_fixed/nova_floating networks
We want to set the MTU of all those networks to the $want_mtu_size var
as we want to allow setting up the custom MTU for all netwroks trougth
the network cookbook.
Not setting it would make the controller tests fail to ping the admin
as the mtu size set for the br-fixed iface would be too low.