capture more logs in supportconfigs
When builds fail it's crucial that we capture enough logs to be
able to debug the problem.  However there are currently two problems:

1) supportconfig defaults to not capturing rotated logs, but those
   might be required sometimes if log rotation occurred during the CI
   run.

2) supportconfig defaults to capturing only the last 500 lines of the
   log files it collects (a restriction which does not apply to the
   logs collected by the SOC plugin, which instead hardcodes the limit
   to 10k lines[0]).  This is a particularly big problem because it
   means we only get the last 500 lines of
   /var/log/{messages,pacemaker.log}, which is generally not enough.

To solve 1), we can use supportconfig's limited but still useful
support for capturing rotated logs[1], which is enabled by adding the
"-l" option.

To solve 2), we can increase the capture limit via the "-I" option.
We don't want to use "-I 0" which would disable the limit, since this
might risk some runaway logger to hog too much disk on ci.suse.de
where these logs will be stored as artifacts.  So default to a limit
of 5000 lines, but leave this configurable.

Unfortunately however as described in [1], currently "-l" already
effectively overrides any "-I" option specified by disabling the
threshold altogether (like as if "-I 0" was specified).  So the
default "-I 5000" added here is ineffective.  However, we hope that at
some point in the future, resolution of [1] might decouple the log
rotation and maximum log lines threshold from both being controlled by
"-l", at which point we'll regain control of the maximum lines
threshold.  In the mean time, err on the side of collecting too much
rather than too little, in order to avoid situations where we cannot
address issues due to insufficient debug.

[0] https://bugzilla.suse.com/show_bug.cgi?id=1015818
[1] https://bugzilla.suse.com/show_bug.cgi?id=1036594