mkcloud: Allow custom crowbar networking mode
Introduce a new 'crowbar_networkingmode' variable which defaults to 'single'.
When set to 'team' ('nics' variable must be > 1 then), mkcloud deploys the cloud
with a bonding network setup. This is useful for testing more complex (and more
realistic) scenarios.
Note: The crowbar node itself has no bonding enabled. Our current mkcloud setup
gives the crowbar node only a single nic (even if 'nics' > 1). That's why this
commit adds an extra entry to the conduit_map for the crowbar node.