mkcloud: support update repos during upgrade process
This PR allows to add packages to the PTF repo during
the upgrade process.
Until now it was only possible to do it before installation
of the initial cloud (UPDATEREPOS variable).
The variable UPDATEREPOS_EXTRA offers something similar for
manually triggered runs.

Background:
Some patches for cloud X break the upgrade process from
cloud X-1. In order to test these properly the patched
package(s) for cloud X need to be added during the upgrade
workflow instead of being added before installation of
cloud X-1.

This PR allows to define the variable UPDATEREPOS_UPGRADE
(its a plus separated list of urls, just like UPDATEREPOS).
The content of these repos will be added to the PTF repo during
the upgrade process. If it is empty nothing will be added.