Wait for testvm deletion
Not waiting for the testvm deletion to finish during testsetup can lead to
problems when the testsetup step is repeated multiple times. Such
problems include failure to delete further resources bound to the VM
(e.g. security groups) or name confilicts because a VM with the same name
is created before the old VM is deleted.