iptables: wait for other process to finish and retry
from iptables manpage:
-w, --wait
Wait for the xtables lock. To prevent multiple instances of the
program from running concurrently, an attempt will be made to obtain an
exclusive lock at launch.  By default, the program will exit if the lock
cannot be obtained.  This option will make the program wait until
the exclusive lock can be obtained.